Sound engine fix
L/R were reversed two times
Modulator not initialized
Wrong type for preceeding values in voice.h

+++ sources/sound_engine/voice.h
@@ -103,7 +103,7 @@ private:
 
     // Save state for resampling
     float _lastFraction;
-    qint32 _valPrec3, _valPrec2, _valPrec1;
+    float _valPrec3, _valPrec2, _valPrec1;
     float _firstVal[3];
 
     // Save state for low pass filter
